+++ src/constants.js
@@ -1,7 +1,7 @@
 'use strict';
 
 module.exports = {
-    SDK_VERSION: '2.3.1',
+    SDK_VERSION: '2.3.2',
     API_VERSION: 'v2',
     RETRIES: 3,
     API_ENDPOINTS: {
@@ -21,5 +21,8 @@ module.exports = {
         DEFAULT: { name: 'AMZN-PAY-RSASSA-PSS', saltLength: 20 },
         V2: { name: 'AMZN-PAY-RSASSA-PSS-V2', saltLength: 32 }
     },
-    ACCOUNT_MANAGEMENT: 'merchantAccounts'
+    ACCOUNT_MANAGEMENT: 'merchantAccounts',
+    DISPUTES: 'disputes',
+    CONTEXT: 'contest',
+    FILES: 'files'
 };